Better builtin (#25)
* split sym_init

* split builtin.c into a subdir

* optimize grammar.y (a little)
